+++ derive/src/lib.rs
@@ -76,6 +76,7 @@ struct StateAttrs {
     flush_event: bool,
     log_flush: bool,
     bevy_state: bool,
+    entity_scope: bool,
     apply_flush: bool,
 }
 
@@ -118,6 +119,7 @@ fn parse_state_attrs(input: &DeriveInput) -> Result<StateAttrs> {
                         "flush_event" => state_attrs.flush_event = true,
                         "log_flush" => state_attrs.log_flush = true,
                         "bevy_state" => state_attrs.bevy_state = true,
+                        "entity_scope" => state_attrs.entity_scope = true,
                         "apply_flush" => state_attrs.apply_flush = true,
                         _ => return Err(Error::new_spanned(ident, "invalid state attribute")),
                     }